Add files via upload
In this Project we are using JWT(Json Web Token) to authenticate and authorize the user.Steps:The user should be giving "username" and "password" and hit the restEndpoint. The request goes to server and authenticate the user, after successful authentication it will return jwt token. Further to access any endpoints the user has to send this token along with request.The server now validate the token that user has send and check for the authorities for this particular user. If he had the authority, server pass his request to the controller and give the response.If the token sent by the user is not valid  or user had no authority for a particular request,the server will simply forbid the user from accessing.